# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub William API",
    "description": "Control-plane and gateway API for Hub William agent pools",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/agent-auth/device": {
      "post": {
        "tags": ["agent auth"],
        "operationId": "start_device_authorization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDeviceAuthorization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Device authorization started; poll with the device code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceAuthorizationStarted"
                }
              }
            }
          },
          "422": {
            "description": "Invalid agent label",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-auth/device/decision": {
      "post": {
        "tags": ["agent auth"],
        "operationId": "decide_device_authorization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideDeviceAuthorization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Authorization approved or denied"
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown, expired, or already-decided code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-auth/device/pending": {
      "get": {
        "tags": ["agent auth"],
        "operationId": "list_pending_authorizations",
        "responses": {
          "200": {
            "description": "Pending authorization requests awaiting this operator",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingDeviceAuthorization"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-auth/device/token": {
      "post": {
        "tags": ["agent auth"],
        "operationId": "poll_device_authorization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollDeviceAuthorization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Current authorization status; the key appears exactly once after approval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceAuthorizationToken"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or expired device code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections": {
      "get": {
        "tags": ["agent connections"],
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "Current user's provider connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentConnection"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/deepseek": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "connect_deepseek",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectDeepseekRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "DeepSeek API key validated and encrypted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid DeepSeek API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "DeepSeek validation unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/start": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Provider authorization started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Provider authorization unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/{connection_id}": {
      "get": {
        "tags": ["agent connections"],
        "operationId": "get_connection",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest provider connection status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["agent connections"],
        "operationId": "disconnect",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Connected account pool deleted"
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/{connection_id}/complete": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "complete_authorization",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteAuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authorization completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "422": {
            "description": "Invalid callback URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Provider rejected the exchange",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent-connections/{connection_id}/refresh": {
      "post": {
        "tags": ["agent connections"],
        "operationId": "refresh_connection",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential refreshed or provider reauthorization started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "401": {
            "description": "Hub login required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Provider refresh unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectDeepseekRequest": {
        "type": "object",
        "required": ["api_key"],
        "properties": {
          "api_key": {
            "type": "string"
          }
        }
      },
...
      "DeviceAuthorizationStarted": {
        "type": "object",
        "required": [
          "device_code",
          "user_code",
          "verification_uri",
          "expires_in_seconds",
          "poll_interval_seconds"
        ],
        "properties": {
          "device_code": {
            "type": "string",
            "description": "Secret the agent polls with. Never shown to the operator."
          },
          "expires_in_seconds": {
            "type": "integer",
            "format": "int64"
          },
          "poll_interval_seconds": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "user_code": {
            "type": "string",
            "description": "Short code the operator types or clicks to approve, e.g. \"XK7P-29QD\"."
          },
          "verification_uri": {
            "type": "string",
            "description": "Where the operator approves the request."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PollDeviceAuthorization": {
        "type": "object",
        "required": ["device_code"],
        "properties": {
          "device_code": {
            "type": "string"
          }
        }
      },
...
      "StartDeviceAuthorization": {
        "type": "object",
        "required": ["agent_label"],
        "properties": {
          "agent_label": {
            "type": "string",
            "description": "Human-readable label the approving operator sees, e.g. \"Claude Code on MacBook\"."
          }
        }
      },
...
    }
  },
...
}